Removed some unnecessary code, added some text formatting
