+++ data_input.cpp
@@ -1,11 +1,89 @@
 #include <string>
 #include <vector>
 #include <iostream>
+#include <fstream>
 #include <algorithm>
 #include <tuple>
 #include "Student.h"
 #include "helper.h"
 #include "data_input.h"
+#include "Exceptions.h"
+
+void ReadDataFromConsole(std::vector<Student>& students) {
+    std::string name, last_name, str;
+    int index = 1, grade, exam_grade, n_grades;
+    char add_another, at_random;
+
+    while (true) {
+        Student s;
+        std::cout << "Enter student's Name: ";
+        std::cin >> s.name;
+        std::cout << "Last name: ";
+        std::cin >> s.last_name;
+
+        std::cout << "\nDo you want to generate student's grades at random? (y/n): ";
+        std::cin >> at_random;
+
+        if (at_random == 'n') {
+            std::cout << "Enter as many student's homework grades as you like. Press ENTER twice to finish." << std::endl;
+
+            std::cin.ignore();
+
+            while (getline(std::cin, str) and !str.empty()) {
+                std::stringstream ss(str);
+                while (ss >> grade)
+                    s.grades.push_back(grade);
+            }
+
+            ReadInt(s.exam_grade, "Enter student's egzam grade: ");
+        }
+        else {
+            ReadInt(n_grades, "Enter a number of homework grades to generate: ");
+            GenerateRandomGrades(s, n_grades);
+        }
+
+        students.push_back(s);
+
+        std::cout << "\nDo you wish to add another student? (y/n): ";
+        std::cin >> add_another;
+
+        if (add_another == 'n') {
+            break;
+        }
+    }
+
+    std::sort(students.begin(), students.end(), CompareStudents);
+}
+
+void ReadDataFromFile(std::vector<Student>& students, std::string filePath) {
+    std::ifstream file(filePath);
+    std::string my_string, token;
+    int pos = 0, grade;
+
+    if (!file) {
+        throw FileNotFound();
+    }
+    
+    if (file.is_open()) {
+        getline(file, my_string);
+        int count = CountWords(my_string);
+
+        while (getline(file, my_string)) {
+            std::stringstream ss(my_string);
+            Student student;
+            ss >> student.name >> student.last_name;
+            for (int i = 0; i < count - 3; i++) {
+                ss >> grade;
+                student.grades.push_back(grade);
+            }
+            ss >> student.exam_grade;
+
+            students.push_back(student);
+        }
+    }
+
+    file.close();
+}
 
 void DataInput() {
     std::vector<Student> students, poor_students, smart_students;
