+++ data_input.h
@@ -1,6 +1,10 @@
 #ifndef DATAINPUT_H
 #define DATAINPUT_H
 
+void ReadDataFromConsole(std::vector<Student>& students);
+
+void ReadDataFromFile(std::vector<Student>& students, std::string filePath);
+
 void DataInput();
 
 #endif
